Add a file containing useful python functions for WRF
Add wrf_utilities.py, a file containing several python functions for WRF
post processing. The most useful are (1) get_wrf_proj, which returns an
object containing the details of the WRF grid and the map projection
used, using a wrfout file or wrfinput file as input (2) wrf_ijll and
wrf_llij, which are used to convert from WRF coordinates to lat/lon and
the opposite, using as input the object returned from get_wrf_proj.